Code coverage back again

+++ tests/bootstrap.php
@@ -16,7 +16,6 @@ use Symfony\Component\Console\Output\ConsoleOutput;
 use Symfony\Component\Console\Output\OutputInterface;
 
 require_once dirname(__DIR__) . '/vendor/autoload.php';
-
 define('APP_NAME', 'EaseFluentPDOTest');
 define('APP_DEBUG', 'true');
 define('DB_CONNECTION', 'sqlite');
@@ -26,8 +25,10 @@ define('DB_DATABASE', __DIR__ . '/test.sqlite');
 define('DB_USERNAME', 'phpunit');
 define('DB_PASSWORD', 'phpunit');
 
-if (file_exists(\Ease\Functions::cfg('DB_DATABASE'))) {
-    unlink(\Ease\Functions::cfg('DB_DATABASE'));
+if (\Ease\Shared::cfg('DB_CONNECTION') == 'sqlite') {
+    if (file_exists(\Ease\Functions::cfg('DB_DATABASE'))) {
+        unlink(\Ease\Functions::cfg('DB_DATABASE'));
+    }
     touch(\Ease\Functions::cfg('DB_DATABASE'));
 } else {
     new DropTable(new Table('test'));
@@ -47,25 +48,27 @@ $manager = new Manager($config, new StringInput(' '), new ConsoleOutput(OutputIn
 $manager->migrate('test');
 $manager->seed('test');
 
-class PDOTester extends \Ease\SQL\PDO {
+class PDOTester extends \Ease\SQL\PDO
+{
 
     use \Ease\SQL\Orm;
 
     public $myTable = 'test';
 
-    public function connect() {
+    public function connect()
+    {
         $this->getPdo();
         parent::connect();
     }
-
 }
 
-class SQLTester extends \Ease\SQL\SQL {
+class SQLTester extends \Ease\SQL\SQL
+{
 
     public $myTable = 'test';
 
-    public function exeQuery($query) {
+    public function exeQuery($query)
+    {
         
     }
-
 }
